requirements.txt
playwright>=1.43.0
requests>=2.31.0
apscheduler>=3.10.4
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
pandas>=2.2.0
plotly>=5.22.0
pydantic>=2.7.0
